refactor: moved config files into config_files dir
