+++ src/apps/backend/modules/config/internals/config_manager.py
@@ -1,6 +1,7 @@
-from typing import Any, Optional
+from typing import Any, Optional, cast
 
-from modules.config.internals.config_file import ConfigFiles
+from modules.config.internals.config_file_loader import ConfigFilesLoader
+from modules.config.types import T
 
 
 class ConfigManager:
@@ -10,7 +11,7 @@ class ConfigManager:
     def __init__(self) -> None:
         self.config_store: dict[str, Any] = {}
 
-    def get(self, key: str, default: Optional[Any] = None) -> Optional[Any]:
+    def get(self, key: str, default: Optional[T] = None) -> Optional[T]:
         value = self._traverse_config(key)
         return value if value is not None else default
 
@@ -18,13 +19,13 @@ class ConfigManager:
         return self._traverse_config(key) is not None
 
     def load_config(self) -> None:
-        ConfigFiles.load()
-        self.config_store = ConfigFiles.get_config_contents()
+        ConfigFilesLoader.load()
+        self.config_store = ConfigFilesLoader.get_config_contents()
 
-    def _traverse_config(self, key: str) -> Optional[Any]:
+    def _traverse_config(self, key: str) -> Optional[T]:
         value = self.config_store
         for k in key.split(self.CONFIG_KEY_SEPARATOR):
             if not isinstance(value, dict) or k not in value:
                 return None  # Indicates key was not found
             value = value[k]
-        return value
+        return cast(T,value)
